# Mutex to prevent concurrency issues