--- a/ContainerHandling/New-NavContainer.ps1
+++ b/ContainerHandling/New-NavContainer.ps1
@@ -397,9 +397,12 @@ function New-NavContainer {
             $MutexName = "img-$($artifactUrl.Split('?')[0])-$imageName"
             $mtx = New-Object System.Threading.Mutex($false, $MutexName)
             Write-Host "Waiting for exclusive access to check if image '$imageName' needs to be built or rebuilt"
-            if (!($mtx.WaitOne())) {
-                throw "Could not obtain exclusive access"
-             }
+            try {
+                $mtx.WaitOne() | Out-Null
+            }
+            catch [System.Threading.AbandonedMutexException] {
+                Write-Host "Other thread terminated without releasing mutex, we can proceed with exclusive access"
+            }
             Write-Host "Got exclusive access to image '$imageName'"
     
             $rebuild = $false