This PR tries to solve some of the issues in #1088 

If several threads downloads artifacts and builds docker images at almost the same time, things goes much slower since they are fighting about resources as CPU, bandwidth, disk space and disk performance.

With a mutex to prevent different processes to do the same thing at the same time, we let the first process finish it's task before the next process checks if this is needed.

My tests shows significant time gain.

### Repro code

Clear bcartifacts.cache and run below in two PS sessions, but change containerName to be different
````
$credential = New-Object pscredential 'super', (ConvertTo-SecureString -String "P@ssword12" -AsPlainText -Force)
$parameters = @{
    "accept_eula"              = $true
    "accept_outdated"          = $true
    "containerName"            = 'test2'
    "artifactUrl"              = "https://bcartifacts.azureedge.net/sandbox/16.3.14085.14363/gb"
    "imageName"                = 'myimage'
    "auth"                     = "NavUserPassword"
    "Credential"               = $credential
    "dns"                      = '8.8.8.8'
    "alwaysPull"               = $true
    "updateHosts"              = $true
    "includeAL"                = $true
    "includeCSide"             = $false
    "doNotExportObjectsToText" = $true
    "shortcuts"                = "none"
    "myScripts"                = @(@{'MainLoop.ps1' = 'while ($true) { start-sleep -seconds 10 }' })
    "doNotCheckHealth"         = $true
    "enableTaskScheduler"      = $false
}

Write-Host "It took $((Measure-Command -Expression { New-NavContainer @parameters }).TotalSeconds) seconds"
```` 


Clear bcartifacts.cache and run below in two PS sessions at the same time
```` 
Measure-Command -Expression {
Download-Artifacts -artifactUrl https://bcartifacts.azureedge.net/sandbox/16.3.14085.14363/us
}
```` 


### Without this change
Download the same artifacts in two processes: 105 seconds
Create new containers with in two processes, same imagename and artifactUrl: 1007s

### After this change
Download the same artifacts in two processes: 55 seconds
Create new containers with in two processes, same imagename and artifactUrl: 863s



ps. There seems to be some whitespace/line break changes as well, it was something VSCode did without showing the diffs. Sorry about that.