@jwikman 2020-07-09
I just realized that it might be needed an additional mutex for the platform, will update with that.