@jwikman 2020-07-09
Changed target branch to dev, you might want to prettify stuff if you decide to merge this.
Added a new mutex for downloading platform