@freddydk 2020-07-09
I haven't worked with mutex - what happens if something goes wrong and the thread holding the mutex crashes?