@jwikman 2020-07-09
> I haven't worked with mutex - what happens if something goes wrong and the thread holding the mutex crashes?

Good question.
I hadn't tested this, I took for granted that it worked as I expected it to (\*embarassed\*)... nowadays an exception is thrown, it wasn't like that once upon a time...

Killing the process having the mutex results in below error in the waiting thread:
`Download-Artifacts : Exception calling "WaitOne" with "0" argument(s): "The wait completed due to an abandoned mutex."`



I'll update to catch and handle the AbandonedMutexException.