@jwikman 2020-07-09
@freddydk Now I've updated with AbandonedMutexException handling. 

In a lot of cases AbandonedMutexException needs to be handled, because something that might be critical just crashed. But in our scenario we know that we will recover that kind of crash by just continuing.

You can easily test out the behavior by running below in several PS sessions and then kill the one that has the mutex (i.e. the one that are downloading) and see what happens with the waiting threads. Clear cache first. :)

```
Measure-Command -Expression {
Download-Artifacts -artifactUrl https://bcartifacts.azureedge.net/sandbox/16.3.14085.14363/us
}

```